refactor: consolidate build logs into build_job_logs_provider by removing legacy build_logs_provider
